// generated/models/UserVo.ts
/* generated using openapi-typescript-codegen -- do no edit */
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type UserVo = {
    createTime?: string;
    id?: number;
    phone?: string;
    userAvatar?: string;
    userName?: string;
    userProfile?: string;
    userRole?: string;
};